+++ components/pm-subscription-handler/pmsh_service/mod/aai_client.py
@@ -120,8 +120,8 @@ def _filter_nf_data(nf_data, nf_filter):
         for nf in nf_data['results']:
             name_identifier = 'pnf-name' if nf['node-type'] == 'pnf' else 'vnf-name'
             orchestration_status = nf['properties'].get('orchestration-status')
-            if nf_filter.is_nf_in_filter(nf['properties'].get(name_identifier)) \
-                    and orchestration_status == 'Active':
+            if nf_filter.is_nf_in_filter(nf['properties'].get(name_identifier),
+                                         orchestration_status):
                 nf_set.add(NetworkFunction(
                     nf_name=nf['properties'].get(name_identifier),
                     orchestration_status=orchestration_status))